create desktop header
