+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/DesktopHeader/index.ts
@@ -1,0 +1,3 @@
+import DesktopHeader from './DesktopHeader';
+
+export default DesktopHeader;
